nrm_backup: Fixup issue with fromDate backup
In case 'fromDate' backup when date was selected with period of first week (in number, the date below 10). That cause invalid date pick from the date algorithm that was exist here.

Old algorithm:
date = current date (in integer)  + 1
monthYear= current year + month (to string)
actualDate = monthYear + date

New algorithm:
date = current date (in integer)  + 1
if date < 10 then
newaffixdate = "0" + date
monthYear= current year + month (to string)
actualDate = monthYear + newaffixdate
else
monthYear= current year + month (to string)
actualDate = monthYear + date
endif

Because of integer use, so when date was picked on 01 -> 09 that was convert to 1 -> 9 on date picker,  but if date was not in range 1 -> 9 then conver to integer doesn't affect the date itself

Issue (with range 1 -> 9):
* 2022036 (With integer conversion)
* 20220306 (Expected result with fix)

Issue (with range (10 -> 30):
* 20220527 (With integer conversion)
* 20220527 (Expected result)

To fix this issue, do another algorithm to detect if date was below than 10, after that include affix "0" as string to append actual date to transform and make sure date can be change to actual format.